chore: Upgrade babel and storybook for various security fixes
