+++ package.json
@@ -12,18 +12,19 @@
     "build-storybook": "build-storybook"
   },
   "devDependencies": {
-    "@storybook/addon-centered": "^5.3.13",
-    "@storybook/react": "^5.3.13",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^8.1.2",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-es2015-destructuring": "^6.23.0",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-flow": "^6.23.0",
-    "babel-preset-react": "^6.24.1",
+    "@babel/cli": "^7.12.10",
+    "@babel/core": "^7.11.1",
+    "@babel/plugin-transform-destructuring": "^7.10.4",
+    "@babel/polyfill": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
+    "@babel/preset-flow": "^7.10.4",
+    "@babel/preset-react": "^7.10.4",
+    "@storybook/addon-actions": "^6.1.15",
+    "@storybook/addon-essentials": "^6.1.15",
+    "@storybook/addon-links": "^6.1.15",
+    "@storybook/react": "^6.1.15",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-styled-components": "^1.12.0",
     "flow-bin": "^0.122.0",
     "react": "^17.0.0",
     "react-dom": "^17.0.0"
