+++ src/index.js
@@ -1,172 +1,85 @@
 // @flow
-import AcademicCapIcon from "./components/AcademicCapIcon";
-import AlignCenterIcon from "./components/AlignCenterIcon";
-import AlignImageCenterIcon from "./components/AlignImageCenterIcon";
-import AlignImageLeftIcon from "./components/AlignImageLeftIcon";
-import AlignImageRightIcon from "./components/AlignImageRightIcon";
-import AlignLeftIcon from "./components/AlignLeftIcon";
-import AlignRightIcon from "./components/AlignRightIcon";
-import AlphabeticalSortIcon from "./components/AlphabeticalSortIcon";
-import ArchiveIcon from "./components/ArchiveIcon";
-import BackIcon from "./components/BackIcon";
-import BeakerIcon from "./components/BeakerIcon";
-import BillingIcon from "./components/BillingIcon";
-import BlockQuoteIcon from "./components/BlockQuoteIcon";
-import BoldIcon from "./components/BoldIcon";
-import BuildingBlocksIcon from "./components/BuildingBlocksIcon";
-import BulletedListIcon from "./components/BulletedListIcon";
-import CheckboxIcon from "./components/CheckboxIcon";
-import CheckmarkIcon from "./components/CheckmarkIcon";
-import CloseIcon from "./components/CloseIcon";
-import CloudIcon from "./components/CloudIcon";
-import CodeIcon from "./components/CodeIcon";
-import CoinsIcon from "./components/CoinsIcon";
-import CollapsedIcon from "./components/CollapsedIcon";
-import CollectionIcon from "./components/CollectionIcon";
-import CommentIcon from "./components/CommentIcon";
-import DocumentIcon from "./components/DocumentIcon";
-import EditIcon from "./components/EditIcon";
-import EmailIcon from "./components/EmailIcon";
-import ExpandedIcon from "./components/ExpandedIcon";
-import EyeIcon from "./components/EyeIcon";
-import GlobeIcon from "./components/GlobeIcon";
-import GoToIcon from "./components/GoToIcon";
-import GroupIcon from "./components/GroupIcon";
-import HashtagIcon from "./components/HashtagIcon";
-import Heading1Icon from "./components/Heading1Icon";
-import Heading2Icon from "./components/Heading2Icon";
-import Heading3Icon from "./components/Heading3Icon";
-import HighlightIcon from "./components/HighlightIcon";
-import HomeIcon from "./components/HomeIcon";
-import HorizontalRuleIcon from "./components/HorizontalRuleIcon";
-import Icon from "./components/Icon";
-import ImageIcon from "./components/ImageIcon";
-import InfoIcon from "./components/InfoIcon";
-import InputIcon from "./components/InputIcon";
-import InsertAboveIcon from "./components/InsertAboveIcon";
-import InsertBelowIcon from "./components/InsertBelowIcon";
-import InsertLeftIcon from "./components/InsertLeftIcon";
-import InsertRightIcon from "./components/InsertRightIcon";
-import ItalicIcon from "./components/ItalicIcon";
-import JournalIcon from "./components/JournalIcon";
-import KeyboardIcon from "./components/KeyboardIcon";
-import LeafIcon from "./components/LeafIcon";
-import LightBulbIcon from "./components/LightBulbIcon";
-import LinkIcon from "./components/LinkIcon";
-import ManualSortIcon from "./components/ManualSortIcon";
-import MenuIcon from "./components/MenuIcon";
-import MoonIcon from "./components/MoonIcon";
-import MoreIcon from "./components/MoreIcon";
-import NewDocumentIcon from "./components/NewDocumentIcon";
-import NextIcon from "./components/NextIcon";
-import NotepadIcon from "./components/NotepadIcon";
-import OpenIcon from "./components/OpenIcon";
-import OrderedListIcon from "./components/OrderedListIcon";
-import PadlockIcon from "./components/PadlockIcon";
-import PaletteIcon from "./components/PaletteIcon";
-import PinIcon from "./components/PinIcon";
-import PlusIcon from "./components/PlusIcon";
-import PrivateCollectionIcon from "./components/PrivateCollectionIcon";
-import ProfileIcon from "./components/ProfileIcon";
-import QuestionMarkIcon from "./components/QuestionMarkIcon";
-import SearchIcon from "./components/SearchIcon";
-import SettingsIcon from "./components/SettingsIcon";
-import ShapesIcon from "./components/ShapesIcon";
-import StarredIcon from "./components/StarredIcon";
-import StrikethroughIcon from "./components/StrikethroughIcon";
-import SunIcon from "./components/SunIcon";
-import TableIcon from "./components/TableIcon";
-import TableOfContentsIcon from "./components/TableOfContentsIcon";
-import TeamIcon from "./components/TeamIcon";
-import TodoListIcon from "./components/TodoListIcon";
-import TrashIcon from "./components/TrashIcon";
-import UserIcon from "./components/UserIcon";
-import VehicleIcon from "./components/VehicleIcon";
-import WarningIcon from "./components/WarningIcon";
-
-module.exports = {
-  AcademicCapIcon,
-  AlignCenterIcon,
-  AlignImageCenterIcon,
-  AlignImageLeftIcon,
-  AlignImageRightIcon,
-  AlignLeftIcon,
-  AlignRightIcon,
-  AlphabeticalSortIcon,
-  ArchiveIcon,
-  BackIcon,
-  BeakerIcon,
-  BillingIcon,
-  BlockQuoteIcon,
-  BoldIcon,
-  BuildingBlocksIcon,
-  BulletedListIcon,
-  CheckboxIcon,
-  CheckmarkIcon,
-  CloseIcon,
-  CloudIcon,
-  CodeIcon,
-  CoinsIcon,
-  CollapsedIcon,
-  CollectionIcon,
-  CommentIcon,
-  DocumentIcon,
-  EditIcon,
-  EmailIcon,
-  ExpandedIcon,
-  EyeIcon,
-  GlobeIcon,
-  GoToIcon,
-  GroupIcon,
-  HashtagIcon,
-  Heading1Icon,
-  Heading2Icon,
-  Heading3Icon,
-  HighlightIcon,
-  HomeIcon,
-  HorizontalRuleIcon,
-  Icon,
-  ImageIcon,
-  InfoIcon,
-  InputIcon,
-  InsertAboveIcon,
-  InsertBelowIcon,
-  InsertLeftIcon,
-  InsertRightIcon,
-  ItalicIcon,
-  JournalIcon,
-  KeyboardIcon,
-  LeafIcon,
-  LightBulbIcon,
-  LinkIcon,
-  ManualSortIcon,
-  MenuIcon,
-  MoonIcon,
-  MoreIcon,
-  NewDocumentIcon,
-  NextIcon,
-  NotepadIcon,
-  OpenIcon,
-  OrderedListIcon,
-  PadlockIcon,
-  PaletteIcon,
-  PinIcon,
-  PlusIcon,
-  PrivateCollectionIcon,
-  ProfileIcon,
-  QuestionMarkIcon,
-  SearchIcon,
-  SettingsIcon,
-  ShapesIcon,
-  StarredIcon,
-  StrikethroughIcon,
-  SunIcon,
-  TableIcon,
-  TableOfContentsIcon,
-  TeamIcon,
-  TodoListIcon,
-  TrashIcon,
-  UserIcon,
-  VehicleIcon,
-  WarningIcon
-};
+export { default as AcademicCapIcon } from "./components/AcademicCapIcon";
+export { default as AlignCenterIcon } from "./components/AlignCenterIcon";
+export { default as AlignImageCenterIcon } from "./components/AlignImageCenterIcon";
+export { default as AlignImageLeftIcon } from "./components/AlignImageLeftIcon";
+export { default as AlignImageRightIcon } from "./components/AlignImageRightIcon";
+export { default as AlignLeftIcon } from "./components/AlignLeftIcon";
+export { default as AlignRightIcon } from "./components/AlignRightIcon";
+export { default as AlphabeticalSortIcon } from "./components/AlphabeticalSortIcon";
+export { default as ArchiveIcon } from "./components/ArchiveIcon";
+export { default as BackIcon } from "./components/BackIcon";
+export { default as BeakerIcon } from "./components/BeakerIcon";
+export { default as BillingIcon } from "./components/BillingIcon";
+export { default as BlockQuoteIcon } from "./components/BlockQuoteIcon";
+export { default as BoldIcon } from "./components/BoldIcon";
+export { default as BuildingBlocksIcon } from "./components/BuildingBlocksIcon";
+export { default as BulletedListIcon } from "./components/BulletedListIcon";
+export { default as CheckboxIcon } from "./components/CheckboxIcon";
+export { default as CheckmarkIcon } from "./components/CheckmarkIcon";
+export { default as CloseIcon } from "./components/CloseIcon";
+export { default as CloudIcon } from "./components/CloudIcon";
+export { default as CodeIcon } from "./components/CodeIcon";
+export { default as CoinsIcon } from "./components/CoinsIcon";
+export { default as CollapsedIcon } from "./components/CollapsedIcon";
+export { default as CollectionIcon } from "./components/CollectionIcon";
+export { default as CommentIcon } from "./components/CommentIcon";
+export { default as DocumentIcon } from "./components/DocumentIcon";
+export { default as EditIcon } from "./components/EditIcon";
+export { default as EmailIcon } from "./components/EmailIcon";
+export { default as ExpandedIcon } from "./components/ExpandedIcon";
+export { default as EyeIcon } from "./components/EyeIcon";
+export { default as GlobeIcon } from "./components/GlobeIcon";
+export { default as GoToIcon } from "./components/GoToIcon";
+export { default as GroupIcon } from "./components/GroupIcon";
+export { default as HashtagIcon } from "./components/HashtagIcon";
+export { default as Heading1Icon } from "./components/Heading1Icon";
+export { default as Heading2Icon } from "./components/Heading2Icon";
+export { default as Heading3Icon } from "./components/Heading3Icon";
+export { default as HighlightIcon } from "./components/HighlightIcon";
+export { default as HomeIcon } from "./components/HomeIcon";
+export { default as HorizontalRuleIcon } from "./components/HorizontalRuleIcon";
+export { default as Icon } from "./components/Icon";
+export { default as ImageIcon } from "./components/ImageIcon";
+export { default as InfoIcon } from "./components/InfoIcon";
+export { default as InputIcon } from "./components/InputIcon";
+export { default as InsertAboveIcon } from "./components/InsertAboveIcon";
+export { default as InsertBelowIcon } from "./components/InsertBelowIcon";
+export { default as InsertLeftIcon } from "./components/InsertLeftIcon";
+export { default as InsertRightIcon } from "./components/InsertRightIcon";
+export { default as ItalicIcon } from "./components/ItalicIcon";
+export { default as JournalIcon } from "./components/JournalIcon";
+export { default as KeyboardIcon } from "./components/KeyboardIcon";
+export { default as LeafIcon } from "./components/LeafIcon";
+export { default as LightBulbIcon } from "./components/LightBulbIcon";
+export { default as LinkIcon } from "./components/LinkIcon";
+export { default as ManualSortIcon } from "./components/ManualSortIcon";
+export { default as MenuIcon } from "./components/MenuIcon";
+export { default as MoonIcon } from "./components/MoonIcon";
+export { default as MoreIcon } from "./components/MoreIcon";
+export { default as NewDocumentIcon } from "./components/NewDocumentIcon";
+export { default as NextIcon } from "./components/NextIcon";
+export { default as NotepadIcon } from "./components/NotepadIcon";
+export { default as OpenIcon } from "./components/OpenIcon";
+export { default as OrderedListIcon } from "./components/OrderedListIcon";
+export { default as PadlockIcon } from "./components/PadlockIcon";
+export { default as PaletteIcon } from "./components/PaletteIcon";
+export { default as PinIcon } from "./components/PinIcon";
+export { default as PlusIcon } from "./components/PlusIcon";
+export { default as PrivateCollectionIcon } from "./components/PrivateCollectionIcon";
+export { default as ProfileIcon } from "./components/ProfileIcon";
+export { default as QuestionMarkIcon } from "./components/QuestionMarkIcon";
+export { default as SearchIcon } from "./components/SearchIcon";
+export { default as SettingsIcon } from "./components/SettingsIcon";
+export { default as ShapesIcon } from "./components/ShapesIcon";
+export { default as StarredIcon } from "./components/StarredIcon";
+export { default as StrikethroughIcon } from "./components/StrikethroughIcon";
+export { default as SunIcon } from "./components/SunIcon";
+export { default as TableIcon } from "./components/TableIcon";
+export { default as TableOfContentsIcon } from "./components/TableOfContentsIcon";
+export { default as TeamIcon } from "./components/TeamIcon";
+export { default as TodoListIcon } from "./components/TodoListIcon";
+export { default as TrashIcon } from "./components/TrashIcon";
+export { default as UserIcon } from "./components/UserIcon";
+export { default as VehicleIcon } from "./components/VehicleIcon";
+export { default as WarningIcon } from "./components/WarningIcon";
